+++ src/Client.php
@@ -6,7 +6,6 @@
 
 namespace Commercetools\Core;
 
-
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Log\LoggerInterface;