+++ src/Model/Common/DiscountValue.php
@@ -5,7 +5,6 @@
 
 namespace Commercetools\Core\Model\Common;
 
-
 /**
  * @package Commercetools\Core\Model\Common
  * @apidoc http://dev.sphere.io/http-api-projects-productDiscounts.html#product-discount-value