+++ src/Model/ProductType/AttributeDefinition.php
@@ -7,6 +7,8 @@ namespace Commercetools\Core\Model\ProductType;
 
 use Commercetools\Core\Model\Common\JsonObject;
 use Commercetools\Core\Model\Common\LocalizedString;
+use Commercetools\Core\Model\Type\FieldDefinition;
+use Commercetools\Core\Model\Type\FieldType;
 
 /**
  * @package Commercetools\Core\Model\ProductType
@@ -26,7 +28,7 @@ use Commercetools\Core\Model\Common\LocalizedString;
  * @method bool getIsSearchable()
  * @method AttributeDefinition setIsSearchable(bool $isSearchable = null)
  */
-class AttributeDefinition extends JsonObject
+class AttributeDefinition extends FieldDefinition
 {
     public function fieldDefinitions()
     {