+++ src/Model/ProductType/MoneyType.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * @author @ct-jensschulze <jens.schulze@commercetools.de>
+ */
+
+namespace Commercetools\Core\Model\ProductType;
+
+/**
+ * @package Commercetools\Core\Model\ProductType
+ * @method string getName()
+ * @method MoneyType setName(string $name = null)
+ */
+class MoneyType extends AttributeType
+{
+    const NAME = 'money';
+
+    public function fieldTypeDefinition()
+    {
+        return [static::TYPE => '\Commercetools\Core\Model\Common\Money'];
+    }
+}