+++ src/Model/ProductType/NestedType.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * @author @ct-jensschulze <jens.schulze@commercetools.de>
+ */
+
+namespace Commercetools\Core\Model\ProductType;
+
+/**
+ * @package Commercetools\Core\Model\ProductType
+ * @method string getName()
+ * @method NestedType setName(string $name = null)
+ */
+class NestedType extends AttributeType
+{
+    const NAME = 'nested';
+
+    public function fieldTypeDefinition()
+    {
+        return [static::TYPE => '\Commercetools\Core\Model\Common\AttributeCollection'];
+    }
+}