+++ src/Request/Carts/Command/CartSetCustomLineItemCustomTypeAction.php
@@ -1,0 +1,48 @@
+<?php
+/**
+ * @author @ct-jensschulze <jens.schulze@commercetools.de>
+ */
+
+namespace Commercetools\Core\Request\Carts\Command;
+
+use Commercetools\Core\Model\Common\Context;
+use Commercetools\Core\Request\CustomField\Command\SetCustomTypeAction;
+use Commercetools\Core\Model\CustomField\FieldContainer;
+
+/**
+ * @package Commercetools\Core\Request\Carts\Command
+ *
+ * @method string getAction()
+ * @method CartSetCustomLineItemCustomTypeAction setAction(string $action = null)
+ * @method string getTypeId()
+ * @method CartSetCustomLineItemCustomTypeAction setTypeId(string $typeId = null)
+ * @method string getTypeKey()
+ * @method CartSetCustomLineItemCustomTypeAction setTypeKey(string $typeKey = null)
+ * @method string getCustomLineItemId()
+ * @method CartSetCustomLineItemCustomTypeAction setCustomLineItemId(string $customLineItemId = null)
+ * @method FieldContainer getFields()
+ * @method CartSetCustomLineItemCustomTypeAction setFields(FieldContainer $fields = null)
+ */
+class CartSetCustomLineItemCustomTypeAction extends SetCustomTypeAction
+{
+    public function fieldDefinitions()
+    {
+        return [
+            'action' => [static::TYPE => 'string'],
+            'typeId' => [static::TYPE => 'string'],
+            'typeKey' => [static::TYPE => 'string'],
+            'customLineItemId' => [static::TYPE => 'string'],
+            'fields' => [static::TYPE => '\Commercetools\Core\Model\CustomField\FieldContainer'],
+        ];
+    }
+
+    /**
+     * @param array $data
+     * @param Context|callable $context
+     */
+    public function __construct(array $data = [], $context = null)
+    {
+        parent::__construct($data, $context);
+        $this->setAction('setCustomLineItemCustomType');
+    }
+}