+++ tests/unit/Model/Common/AttributeCollectionTest.php
@@ -5,7 +5,6 @@
 
 namespace Commercetools\Core\Model\Common;
 
-
 use Commercetools\Core\Model\ProductType\AttributeDefinition;
 use Commercetools\Core\Model\ProductType\AttributeDefinitionCollection;
 use Commercetools\Core\Model\ProductType\AttributeType;
@@ -15,9 +14,7 @@ class AttributeCollectionTest extends \PHPUnit_Framework_TestCase
     public function testIndex()
     {
         $collection = AttributeCollection::fromArray([
-            [
-                'name' => 'test'
-            ]
+            ['name' => 'test']
         ]);
 
         $this->assertInstanceOf('\Commercetools\Core\Model\Common\Attribute', $collection->getByName('test'));
@@ -75,9 +72,6 @@ class AttributeCollectionTest extends \PHPUnit_Framework_TestCase
 
         $t = $collection->getByName('test-definition-enum');
 
-        $fields = $t->fieldDefinitions();
-        $this->assertSame('\Commercetools\Core\Model\Common\Enum', $fields[Attribute::PROP_VALUE][JsonObject::TYPE]);
-        $this->assertNull($fields[Attribute::PROP_VALUE][JsonObject::DECORATOR]);
-        $this->assertNull($fields[Attribute::PROP_VALUE][JsonObject::ELEMENT_TYPE]);
+        $this->assertInstanceOf('\Commercetools\Core\Model\Common\Enum', $t->getValue());
     }
 }